refactor: apply /simplify cleanups on dspark diff.
- Dedup DSpark SWA window formula: drop dflash local wrapper and call
  layer::deepseek_v4_ori_window_left directly.
- Drop cached speculative_position_labels_ in WorkerService; inline
  std::to_string at the sole call site.
- Precompute an unordered_map slot lookup in AuxHiddenCapture so
  should_capture and capture_layer stop scanning layers_to_capture per
  layer per forward.
- Tighten DFlashWorkerImpl::draft_sas_mode() to CHECK draft_impl_ (only
  reached after init).
- Fold DSparkMarkovHead::initialize into its constructor; drop the
  two-phase init from DeepseekV4DSpark and DSparkQwen3.
- Cache dspark SWA layer/cache/block_size in DeepseekV4ModelImpl so
  build_dspark_swa_metadata indexes directly instead of scanning
  caches_info_ per forward.
- Wrap configure_deepseek_v4_dspark_args in #if USE_NPU and drop the
  unreachable non-NPU LOG(FATAL) branch.
- Drop redundant .to(torch::kLong) on already-kLong block_columns before
  block_table.gather in build_dspark_swa_indices.

+++ xllm/core/layers/npu_torch/deepseek_sparse_attention.cpp
@@ -532,8 +532,7 @@ torch::Tensor build_dspark_swa_indices(const torch::Tensor& block_table,
   torch::Tensor block_columns =
       (positions / cache_block_size)
           .clamp(/*min=*/0, /*max=*/block_table.size(1) - 1);
-  torch::Tensor block_ids =
-      block_table.gather(/*dim=*/1, block_columns.to(torch::kLong));
+  torch::Tensor block_ids = block_table.gather(/*dim=*/1, block_columns);
   torch::Tensor slot_ids =
       block_ids * cache_block_size + positions.remainder(cache_block_size);
   slot_ids = torch::where(valid, slot_ids, torch::full_like(slot_ids, -1));
